create user

+++ server/api/comments/delete-comment.ts
@@ -5,6 +5,13 @@ import { comments } from "~/server/database/schema";
 export default defineEventHandler(async (event) => {
   const body = await readBody(event);
 
+  if (!body?.id) {
+    throw createError({
+      statusCode: 422,
+      message: "ID отзыва отсутствует",
+    });
+  }
+
   const result = await db.delete(comments).where(eq(comments.id, body.id));
 
   return result;
